+++ src/types/symbols/index.ts
@@ -3,3 +3,6 @@ export const Schema = Symbol.for('schema');
 export const QueryPath = Symbol.for('queryPath');
 
 export const StepPrint = Symbol.for('stepPrint');
+
+export const EdgeType = Symbol.for('edgeType');
+export const ParentFieldSchema = Symbol.for('parentFieldSchema');