+++ changelog.md
@@ -2,10 +2,13 @@
 
 📝 following beta format X.Y.Z where Y = breaking change and Z = feature and fix. Later => FAIL.FEATURE.FIX
 
-## 0.8.12(2023-02-16)
+## 0.9.0(2023-02-16)
 
 - Feat: Non linear pipeline
+- Feat: Replace is now a valid operation
+- Breaking change: Should use $thing instead of $entity and $relation;
 - Refacto: Preparing for multiple DBs
+- Chore: Immer downgraded to 9.0.21 (until <https://github.com/immerjs/immer/issues/1087> is fixed)
 
 ## 0.8.10(2023-02-14)